requirements.txt
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-RESTful==0.3.7
gunicorn==19.7.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-gnupg==0.4.5
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0
